requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-decouple==3.8
mysqlclient==2.2.0
redis==5.0.1
celery==5.3.4
django-filter==23.3
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-tables2==2.6.0
django-widget-tweaks==1.5.0
django-import-export==3.3.4
django-activity-stream==2.0.0
django-notifications-hq==1.8.3
django-simple-history==3.4.0
django-guardian==2.4.0
django-ratelimit==4.1.0
django-cacheops==8.0.0
django-redis==5.4.0
django-storages==1.14.2
boto3==1.34.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
